feature(agent-router): variable names changed to meaningful ones, router flow added, sanitization prompt fixed
