+++ .idea/workspace.xml
@@ -4,17 +4,15 @@
     <option name="autoReloadType" value="SELECTIVE" />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="853118d8-f694-45d4-a927-df31f718b200" name="Changes" comment="schema added">
-      <change beforePath="$PROJECT_DIR$/.idea/vcs.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/vcs.xml" afterDir="false" />
+    <list default="true" id="853118d8-f694-45d4-a927-df31f718b200" name="Changes" comment="">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/agents/smolagents_agents/agent.py" beforeDir="false" afterPath="$PROJECT_DIR$/agents/smolagents_agents/agent.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/agents/smolagents_agents/research_tools.py" beforeDir="false" />
       <change beforePath="$PROJECT_DIR$/agents/smolagents_agents/user_tools.py" beforeDir="false" afterPath="$PROJECT_DIR$/agents/smolagents_agents/user_tools.py" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/controller/chatbot.py" beforeDir="false" afterPath="$PROJECT_DIR$/controller/chatbot.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/repository/db.py" beforeDir="false" afterPath="$PROJECT_DIR$/repository/db.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/routers/router.py" beforeDir="false" afterPath="$PROJECT_DIR$/routers/router.py" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/services/agent_service.py" beforeDir="false" afterPath="$PROJECT_DIR$/services/agent_service.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/services/ai_helper_service.py" beforeDir="false" afterPath="$PROJECT_DIR$/services/ai_helper_service.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/services/user_service.py" beforeDir="false" afterPath="$PROJECT_DIR$/services/user_service.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/services/ai_helper_service.py" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/services/helper_service.py" beforeDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -164,7 +162,7 @@
       <workItem from="1740913152925" duration="41000" />
       <workItem from="1740913201813" duration="5209000" />
       <workItem from="1740988536422" duration="9500000" />
-      <workItem from="1741014037230" duration="4771000" />
+      <workItem from="1741014037230" duration="9059000" />
     </task>
     <task id="LOCAL-00001" summary="agents added">
       <option name="closed" value="true" />
@@ -222,7 +220,15 @@
       <option name="project" value="LOCAL" />
       <updated>1740989404903</updated>
     </task>
-    <option name="localTasksCounter" value="8" />
+    <task id="LOCAL-00008" summary="stemma prompt optimized">
+      <option name="closed" value="true" />
+      <created>1741019238203</created>
+      <option name="number" value="00008" />
+      <option name="presentableId" value="LOCAL-00008" />
+      <option name="project" value="LOCAL" />
+      <updated>1741019238203</updated>
+    </task>
+    <option name="localTasksCounter" value="9" />
     <servers />
   </component>
   <component name="TypeScriptGeneratedFilesManager">
@@ -232,6 +238,7 @@
     <MESSAGE value="agents added" />
     <MESSAGE value="agentic workflow added" />
     <MESSAGE value="schema added" />
-    <option name="LAST_COMMIT_MESSAGE" value="schema added" />
+    <MESSAGE value="stemma prompt optimized" />
+    <option name="LAST_COMMIT_MESSAGE" value="stemma prompt optimized" />
   </component>
 </project>
