+++ agents/smolagents_agents/agent.py
@@ -2,7 +2,6 @@ from smolagents import CodeAgent
 
 from agents.smolagents_agents.agent_helper import get_agent_model
 from agents.smolagents_agents.chat_tools import send_chat_message
-from agents.smolagents_agents.research_tools import researcher
 from agents.smolagents_agents.user_tools import get_premium_status, get_trust_score_by_id, get_user_info_by_id, \
     get_user_info_by_phone, block_number_by_id, get_trust_score_phone
 
@@ -19,7 +18,6 @@ def get_agent():
             get_user_info_by_id,
             get_user_info_by_phone,
             block_number_by_id,
-            researcher,
             send_chat_message
         ],
         additional_authorized_imports=[
