--- services/helper_service.py
@@ -1,15 +1,0 @@
-import re, hashlib
-
-
-def is_valid_email(email):
-    return re.match(r"[^@]+@[^@]+\.[^@]+", email) is not None
-
-
-def is_valid_age(age):
-    return 0 < age < 120
-
-
-def generate_hash(input_string: str) -> str:
-    encoded_string = input_string.encode()
-    hash_object = hashlib.sha256(encoded_string)
-    return hash_object.hexdigest()
